+++ scripts/clean.ts
@@ -1,10 +1,10 @@
 import path from 'node:path'
-import execa from 'execa'
+import { execaCommandSync } from 'execa'
 import { logger } from '@ombro/logger'
 
 export function cleanDeps() {
   logger.startLoading('Cleaning dependencies...')
-  execa.commandSync('pnpm -wr exec -- rimraf node_modules', {
+  execaCommandSync('pnpm -wr exec -- rimraf node_modules', {
     stdio: 'inherit',
     cwd: path.resolve(__dirname, '../'),
   })
@@ -14,7 +14,7 @@ export function cleanDeps() {
 
 export function cleanBuild() {
   logger.startLoading('Cleaning build...')
-  execa.commandSync('pnpm -r exec -- rimraf dist coverage *.tsbuildinfo', {
+  execaCommandSync('pnpm -r exec -- rimraf dist coverage *.tsbuildinfo', {
     stdio: 'inherit',
     cwd: path.resolve(__dirname, '../'),
   })