+++ packages/typescript/vue3-vben-admin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vrn-deco/boilerplate-typescript-vue3-vben-admin
 
+## 2.8.2
+
+### Patch Changes
+
+- ESMify
+- Updated dependencies
+  - @vrn-deco/boilerplate-preset-npm@1.1.4
+
 ## 2.8.1
 
 ### Patch Changes