+++ presets/npm/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @vrn-deco/boilerplate-preset-npm
 
+## 1.1.4
+
+### Patch Changes
+
+- ESMify
+- Updated dependencies
+  - @vrn-deco/boilerplate-preset-base@1.3.3
+  - @vrn-deco/boilerplate-protocol@1.3.1
+
 ## 1.1.3
 
 ### Patch Changes