+++ presets/pip/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @vrn-deco/boilerplate-preset-pip
 
+## 0.0.6
+
+### Patch Changes
+
+- ESMify
+- Updated dependencies
+  - @vrn-deco/boilerplate-preset-base@1.3.3
+  - @vrn-deco/boilerplate-protocol@1.3.1
+
 ## 0.0.5
 
 ### Patch Changes